pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "youtube-news-agent"
version = "0.1.0"
description = "A simple YouTube news digest generator."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["youtube_news_agent*"]
